refactor(compiler_utils): Move two methods out of compiler_utils that are only used in a single place.
PiperOrigin-RevId: 246061692
