+++ _tests/test/compiler/ast_template_parser_test.dart
@@ -7,7 +7,6 @@ import 'package:test/test.dart';
 import 'package:_tests/test_util.dart';
 import 'package:angular/src/compiler/analyzed_class.dart';
 import 'package:angular/src/compiler/compile_metadata.dart';
-import 'package:angular/src/compiler/compiler_utils.dart';
 import 'package:angular/src/compiler/expression_parser/lexer.dart';
 import 'package:angular/src/compiler/expression_parser/parser.dart';
 import 'package:angular/src/compiler/identifiers.dart'
@@ -18,6 +17,7 @@ import 'package:angular/src/compiler/schema/element_schema_registry.dart'
     show ElementSchemaRegistry;
 import 'package:angular/src/compiler/template_ast.dart';
 import 'package:angular/src/compiler/template_parser/ast_template_parser.dart';
+import 'package:angular/src/facade/lang.dart' show jsSplit;
 import 'package:angular_compiler/cli.dart';
 
 import 'schema_registry_mock.dart' show MockSchemaRegistry;
@@ -2362,4 +2362,13 @@ CompileDirectiveMetadata createCompileDirectiveMetadata({
       queries: queries,
       template: template ?? CompileTemplateMetadata(),
       analyzedClass: AnalyzedClass(null));
+}
+
+List<String> splitAtColon(String input, List<String> defaultValues) {
+  var parts = jsSplit(input.trim(), RegExp(r'\s*:\s*'));
+  if (parts.length > 1) {
+    return parts;
+  } else {
+    return defaultValues;
+  }
 }
