+++ angular/lib/src/compiler/compiler_utils.dart
@@ -1,21 +1,6 @@
-import "package:angular/src/facade/lang.dart" show jsSplit;
-
 import 'compile_metadata.dart';
 
 const moduleSuffix = ".dart";
-
-List<String> splitAtColon(String input, List<String> defaultValues) {
-  var parts = jsSplit(input.trim(), RegExp(r'\s*:\s*'));
-  if (parts.length > 1) {
-    return parts;
-  } else {
-    return defaultValues;
-  }
-}
-
-String sanitizeIdentifier(Object name) {
-  return name.toString().replaceAll(RegExp(r'\W'), "_");
-}
 
 String templateModuleUrl(CompileTypeMetadata type) {
   var moduleUrl = type.moduleUrl;
